+++ src/spatial_metrics.rs
@@ -1,16 +1,16 @@
-use libm::acosf;
+use libm::acos;
 use na::{DMatrix, MatrixView};
 use num_traits::Pow;
-use std::{f32::consts::PI, f32::EPSILON};
+use std::{f64::consts::PI, f64::EPSILON};
 
-pub fn relative_difference(x: f32, y: f32) -> f32 {
+pub fn relative_difference(x: f64, y: f64) -> f64 {
     1. - (x - y).abs() / (x.abs() + y.abs() + EPSILON)
 }
 
 pub fn iter_relative_difference<'a, T: na::Dim>(
-    x: &'a MatrixView<f32, T, T>,
-    y: &'a MatrixView<f32, T, T>,
-) -> DMatrix<f32> {
+    x: &'a MatrixView<f64, T, T>,
+    y: &'a MatrixView<f64, T, T>,
+) -> DMatrix<f64> {
     let nrows = x.nrows();
     let ncols = x.ncols();
     let mut result = DMatrix::zeros(nrows, ncols);
@@ -23,10 +23,10 @@ pub fn iter_relative_difference<'a, T: na::Dim>(
 }
 
 pub fn covariance_differences<'a, T: na::Dim>(
-    x: &'a MatrixView<f32, T, T>,
-    y: &'a MatrixView<f32, T, T>,
-    z: &'a MatrixView<f32, T, T>,
-) -> DMatrix<f32> {
+    x: &'a MatrixView<f64, T, T>,
+    y: &'a MatrixView<f64, T, T>,
+    z: &'a MatrixView<f64, T, T>,
+) -> DMatrix<f64> {
     let nrows = x.nrows();
     let ncols = x.ncols();
     let mut result = DMatrix::zeros(nrows, ncols);
@@ -41,43 +41,43 @@ pub fn covariance_differences<'a, T: na::Dim>(
 }
 
 pub fn textural_variance_sum<'a, T: na::Dim>(
-    x: &'a MatrixView<f32, T, T>,
-    y: &'a MatrixView<f32, T, T>,
-) -> DMatrix<f32> {
+    x: &'a MatrixView<f64, T, T>,
+    y: &'a MatrixView<f64, T, T>,
+) -> DMatrix<f64> {
     let nrows = x.nrows();
     let mut result = DMatrix::zeros(nrows, 1);
     for i in 0..nrows {
-        let x_sum: f32 = x.row(i).sum();
-        let y_sum: f32 = y.row(i).sum();
+        let x_sum: f64 = x.row(i).sum();
+        let y_sum: f64 = y.row(i).sum();
         result[(i, 0)] = relative_difference(x_sum, y_sum);
     }
     result
 }
 
 pub fn omnivariance_differences<'a, T: na::Dim>(
-    x: &'a MatrixView<f32, T, T>,
-    y: &'a MatrixView<f32, T, T>,
-) -> DMatrix<f32> {
+    x: &'a MatrixView<f64, T, T>,
+    y: &'a MatrixView<f64, T, T>,
+) -> DMatrix<f64> {
     let nrows = x.nrows();
     let mut result = DMatrix::zeros(nrows, 1);
     for i in 0..nrows {
-        let x_prod: f32 = x.row(i).product();
-        let y_prod: f32 = y.row(i).product();
+        let x_prod: f64 = x.row(i).product();
+        let y_prod: f64 = y.row(i).product();
         result[(i, 0)] = relative_difference(x_prod.cbrt(), y_prod.cbrt());
     }
     result
 }
 
 pub fn entropy<'a, T: na::Dim>(
-    x: &'a MatrixView<f32, T, T>,
-    y: &'a MatrixView<f32, T, T>,
-) -> DMatrix<f32> {
+    x: &'a MatrixView<f64, T, T>,
+    y: &'a MatrixView<f64, T, T>,
+) -> DMatrix<f64> {
     let nrows = x.nrows();
     let ncols = x.ncols();
     let mut result = DMatrix::zeros(nrows, 1);
     for i in 0..nrows {
-        let mut x_entropy: f32 = 0.;
-        let mut y_entropy: f32 = 0.;
+        let mut x_entropy: f64 = 0.;
+        let mut y_entropy: f64 = 0.;
         for j in 0..ncols {
             x_entropy += x[(i, j)] * (x[(i, j)] + EPSILON).ln();
             y_entropy += y[(i, j)] * (y[(i, j)] + EPSILON).ln();
@@ -88,14 +88,14 @@ pub fn entropy<'a, T: na::Dim>(
 }
 
 pub fn euclidean_distances<'a, T: na::Dim>(
-    x: &'a MatrixView<f32, T, T>,
-    y: &'a MatrixView<f32, T, T>,
-) -> DMatrix<f32> {
+    x: &'a MatrixView<f64, T, T>,
+    y: &'a MatrixView<f64, T, T>,
+) -> DMatrix<f64> {
     let nrows = x.nrows();
     let ncols = x.ncols();
     let mut result = DMatrix::zeros(nrows, 1);
     for i in 0..nrows {
-        let mut dists: f32 = 0.;
+        let mut dists: f64 = 0.;
         for j in 0..ncols {
             dists += (x[(i, j)] - y[(i, j)]).pow(2);
         }
@@ -105,10 +105,10 @@ pub fn euclidean_distances<'a, T: na::Dim>(
 }
 
 pub fn vector_projected_distances<'a, T: na::Dim>(
-    x: &'a MatrixView<f32, T, T>,
-    y: &'a MatrixView<f32, T, T>,
+    x: &'a MatrixView<f64, T, T>,
+    y: &'a MatrixView<f64, T, T>,
     col: usize,
-) -> DMatrix<f32> {
+) -> DMatrix<f64> {
     let ncols = 1;
     let nrows = x.nrows();
     let mut result = DMatrix::zeros(nrows, ncols);
@@ -118,7 +118,7 @@ pub fn vector_projected_distances<'a, T: na::Dim>(
     result
 }
 
-pub fn point_projected_distances<'a, T: na::Dim>(x: &'a MatrixView<f32, T, T>) -> DMatrix<f32> {
+pub fn point_projected_distances<'a, T: na::Dim>(x: &'a MatrixView<f64, T, T>) -> DMatrix<f64> {
     let ncols = 2;
     let nrows = x.nrows();
     let mut result = DMatrix::zeros(nrows, ncols);
@@ -129,12 +129,12 @@ pub fn point_projected_distances<'a, T: na::Dim>(x: &'a MatrixView<f32, T, T>) -
     result
 }
 
-pub fn point_to_centroid_distances<'a, T: na::Dim>(x: &'a MatrixView<f32, T, T>) -> DMatrix<f32> {
+pub fn point_to_centroid_distances<'a, T: na::Dim>(x: &'a MatrixView<f64, T, T>) -> DMatrix<f64> {
     let ncols = 1;
     let nrows = x.nrows();
     let mut result = DMatrix::zeros(nrows, ncols);
     for i in 0..nrows {
-        let mut dists: f32 = 0.;
+        let mut dists: f64 = 0.;
         for j in 0..x.ncols() {
             dists += x[(i, j)].pow(2);
         }
@@ -144,11 +144,11 @@ pub fn point_to_centroid_distances<'a, T: na::Dim>(x: &'a MatrixView<f32, T, T>)
 }
 
 pub fn anisotropy_planarity_linearity<'a, T: na::Dim>(
-    x: &'a MatrixView<f32, T, T>,
-    y: &'a MatrixView<f32, T, T>,
+    x: &'a MatrixView<f64, T, T>,
+    y: &'a MatrixView<f64, T, T>,
     col1: usize,
     col2: usize,
-) -> DMatrix<f32> {
+) -> DMatrix<f64> {
     let nrows = x.nrows();
     let mut result = DMatrix::zeros(nrows, 1);
     for i in 0..nrows {
@@ -160,14 +160,14 @@ pub fn anisotropy_planarity_linearity<'a, T: na::Dim>(
 }
 
 pub fn surface_variation<'a, T: na::Dim>(
-    x: &'a MatrixView<f32, T, T>,
-    y: &'a MatrixView<f32, T, T>,
-) -> DMatrix<f32> {
+    x: &'a MatrixView<f64, T, T>,
+    y: &'a MatrixView<f64, T, T>,
+) -> DMatrix<f64> {
     let nrows = x.nrows();
     let mut result = DMatrix::zeros(nrows, 1);
     for i in 0..nrows {
-        let x_sum: f32 = x.row(i).sum();
-        let y_sum: f32 = y.row(i).sum();
+        let x_sum: f64 = x.row(i).sum();
+        let y_sum: f64 = y.row(i).sum();
         result[(i, 0)] =
             relative_difference(x[(i, 2)] / (x_sum + EPSILON), y[(i, 2)] / (y_sum + EPSILON));
     }
@@ -175,9 +175,9 @@ pub fn surface_variation<'a, T: na::Dim>(
 }
 
 pub fn sphericity<'a, T: na::Dim>(
-    x: &'a MatrixView<f32, T, T>,
-    y: &'a MatrixView<f32, T, T>,
-) -> DMatrix<f32> {
+    x: &'a MatrixView<f64, T, T>,
+    y: &'a MatrixView<f64, T, T>,
+) -> DMatrix<f64> {
     let nrows = x.nrows();
     let mut result = DMatrix::zeros(nrows, 1);
     for i in 0..nrows {
@@ -189,20 +189,20 @@ pub fn sphericity<'a, T: na::Dim>(
     result
 }
 
-pub fn angular_similarity<'a, T: na::Dim>(x: &'a MatrixView<f32, T, T>) -> DMatrix<f32> {
+pub fn angular_similarity<'a, T: na::Dim>(x: &'a MatrixView<f64, T, T>) -> DMatrix<f64> {
     let nrows = x.nrows();
     let mut result = DMatrix::zeros(nrows, 1);
     for i in 0..nrows {
         let numerator = x[(i, 1)];
         let (mut a, mut b, mut c) = (x[(i, 0)], x[(i, 1)], x[(i, 2)]);
         (a, b, c) = (a.pow(2), b.pow(2), c.pow(2));
-        let denominator: f32 = (a + b + c).sqrt() + EPSILON;
-        result[(i, 0)] = 1. - 2. * acosf((numerator / denominator).abs()) / PI;
+        let denominator: f64 = (a + b + c).sqrt() + EPSILON;
+        result[(i, 0)] = 1. - 2. * acos((numerator / denominator).abs()) / PI;
     }
     result
 }
 
-pub fn parallelity<'a, T: na::Dim>(x: &'a MatrixView<f32, T, T>, col: usize) -> DMatrix<f32> {
+pub fn parallelity<'a, T: na::Dim>(x: &'a MatrixView<f64, T, T>, col: usize) -> DMatrix<f64> {
     let nrows = x.nrows();
     let mut result = DMatrix::zeros(nrows, 1);
     for (i, num) in x.column(col).iter().enumerate() {
